Extract audio player to protocol
